windows: fix image corruption from 16-bit channel truncation
writeImage packed pixels with uint8(r) on the values returned by
color.Color.RGBA(), which are 16-bit (0-0xffff). Taking the low byte
is correct only for 8-bit-sourced images, where the high and low bytes
are equal. For 16-bit images it stores garbage.

This is exactly what the reported pipeline hits: decoding a JPEG yields
an *image.YCbCr, and png.Encode emits 16-bit truecolor for that color
model, so png.Decode hands back an *image.RGBA64. Every pixel written to
the clipboard was then corrupted, while the same bytes saved to a file
stayed correct.

Take the high byte (uint8(r >> 8)) instead. The existing round-trip test
missed this because its fixture is 8-bit and both write and read applied
the same truncation, cancelling out.

The pure image->DIB conversion is extracted into an untagged dib.go so it
can be unit tested on any platform; the new tests cover a 16-bit image and
the full JPEG->PNG pipeline, and fail without this fix.

Fixes #48
